Rename "utils" -> "liveblocks-helpers"
This is done to stress the nature of these "utilities": they're very
much Liveblocks-specific, and should ideally be organized into
their more related modules. Right now they're a bit of a potpourri of
random bits of functionality.
